Optimized array, list and map for embedded systems

+++ src/data/array.h
@@ -15,7 +15,6 @@
 extern "C" {
 #endif
 
-#include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
 
@@ -35,7 +34,7 @@ typedef enum {
 typedef struct XArrayData {
     void* pData;
     size_t nSize;
-    uint64_t nKey;
+    uint32_t nKey;
 } xarray_data_t;
 
 typedef int(*xarray_comparator_t)(const void*, const void*, void*);
@@ -51,7 +50,7 @@ typedef struct XArray_ {
     size_t nUsed;
 } xarray_t;
 
-xarray_data_t *XArray_NewData(void *pData, size_t nSize, uint64_t nKey);
+xarray_data_t *XArray_NewData(void *pData, size_t nSize, uint32_t nKey);
 void XArray_FreeData(xarray_data_t *pArrData);
 void XArray_ClearData(xarray_t *pArr, xarray_data_t *pArrData);
 
@@ -65,11 +64,11 @@ void XArray_Free(xarray_t **ppArr);
 int XArray_Add(xarray_t *pArr, xarray_data_t *pNewData);
 int XArray_AddData(xarray_t *pArr, void* pData, size_t nSize);
 int XArray_PushData(xarray_t *pArr, void *pData, size_t nSize);
-int XArray_AddDataKey(xarray_t *pArr, void* pData, size_t nSize, uint64_t nKey);
+int XArray_AddDataKey(xarray_t *pArr, void* pData, size_t nSize, uint32_t nKey);
 void* XArray_GetDataOr(xarray_t *pArr, size_t nIndex, void *pRet);
 void* XArray_GetData(xarray_t *pArr, size_t nIndex);
 size_t XArray_GetSize(xarray_t *pArr, size_t nIndex);
-uint64_t XArray_GetKey(xarray_t *pArr, size_t nIndex);
+uint32_t XArray_GetKey(xarray_t *pArr, size_t nIndex);
 uint8_t XArray_Contains(xarray_t *pArr, size_t nIndex);
 
 xarray_data_t* XArray_Remove(xarray_t *pArr, size_t nIndex);
@@ -81,10 +80,10 @@ void XArray_BubbleSort(xarray_t *pArr, xarray_comparator_t compare, void *pCtx);
 void XArray_QuickSort(xarray_t *pArr, xarray_comparator_t compare, void *pCtx, int nStart, int nFinish);
 void XArray_SortBy(xarray_t *pArr, int nSortBy);
 
-int XArray_SentinelSearch(xarray_t *pArr, uint64_t nKey);
-int XArray_LinearSearch(xarray_t *pArr, uint64_t nKey);
-int XArray_DoubleSearch(xarray_t *pArr, uint64_t nKey);
-int XArray_BinarySearch(xarray_t *pArr, uint64_t nKey);
+int XArray_SentinelSearch(xarray_t *pArr, uint32_t nKey);
+int XArray_LinearSearch(xarray_t *pArr, uint32_t nKey);
+int XArray_DoubleSearch(xarray_t *pArr, uint32_t nKey);
+int XArray_BinarySearch(xarray_t *pArr, uint32_t nKey);
 
 xarray_data_t* XArray_Get(xarray_t *pArr, size_t nIndex);
 xarray_data_t* XArray_Set(xarray_t *pArr, size_t nIndex, xarray_data_t *pNewData);
